Moved measureTiming to seperate file, in order to replicate the structure of DCR's timing functionality
